requirements/local.txt
celery
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-celery-email
django-environ
flower
pillow
psycopg2-binary
redis
watchdog